requirements.txt
streamlit
langchain
langchain-community
unstructured
openai
faiss-cpu
python-dotenv
requests